BUG/MINOR: hlua: Rely on CF_EOI to detect end of message in HTTP applets
applet:getline() and applet:receive() functions for HTTP applets must rely
on the channel flags to detect the end of the message and not on HTX
flags. It means CF_EOI must be used instead of HTX_FL_EOM.

It is important because the HTX flag is transient. Because there is no flag
on HTTP applets to save the info, it is not reliable. However CF_EOI once
set is never removed. So it is safer to rely on it. Otherwise, the call to
these functions hang.

This patch must be backported as far as 2.4.

(cherry picked from commit 4a20972a954b2af0ec75b21e32d2d56ae97a8493)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 4109605dc9f0c3ccbb273253952eb2ef1c1dea6a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>