MINOR: debug: report applet pointer and handler in crashes when known
When an appctx is found looping over itself, we report a number of info
but not the pointers to the definition nor the handler, which can be quite
handy in some cases. Let's add them and try to decode the symbol.

(cherry picked from commit 714900a3c967eae7ff8d528239e7e9d02fac5df9)
Signed-off-by: Willy Tarreau <w@1wt.eu>