BUG/MINOR: mux-h1: fix the "show fd" dest buffer for the subscriber
Commit 150c4f8b7 ("MINOR: mux-h1: make the "show fd" helper also decode
the h1s subscriber when known") improved the output of "show fd" for the
H1 mux, but the output is sent to the trash buffer instead of the msg
argument. It turns out that this has no effect right now as the caller
passes the trash but this is risky.

This should be backported to 2.4.

(cherry picked from commit 9b6a187e26cd0b25429ab79737b84c41e939bbcc)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit d6c011021efa3d5c21ff1666cb8344206b6f8461)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>