BUG/MINOR: mux-fcgi: fix the "show fd" dest buffer for the subscriber
Commit 1776ffb97 ("MINOR: mux-fcgi: make the "show fd" helper also decode
the fstrm subscriber when known") improved the output of "show fd" for the
FCGI mux, but the output is sent to the trash buffer instead of the msg
argument. It turns out that this has no effect right now as the caller
passes the trash but this is risky.

This should be backported to 2.4.

(cherry picked from commit 410546145b58adc035c357fb89163ced4fb84829)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit b716db161d36f02ad6ded6865a336efa1d4ae913)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>