BUG/MINOR: h1: Support headers case adjustment for TCP proxies
On frontend side, "h1-case-adjust-bogus-client" option is now supported in
TCP mode. It is important to be able to adjust the case of response headers
when a connection is routed to an HTTP backend. In this case, the client
connection is upgraded to H1.

On backend side, "h1-case-adjust-bogus-server" option is now also supported
in TCP mode to be able to perform HTTP health-checks with a case adjustment
of the request headers.

This patch should be backported as far as 2.0.

(cherry picked from commit a9e934bbd189b7d85af454f03ec1ade692154fd5)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit eff57f2476766fdabb9c5c2638f322abd751cafe)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>