REORG: mux-quic: export HTTP related function in a dedicated file
Extract function dealing with HTX outside of MUX QUIC. For the moment,
only rcv_buf stream operation is concerned.

The main objective is to be able to support both TCP and HTTP proxy mode
with a common base and add specialized modules on top of it.

This should be backported up to 2.6.

(cherry picked from commit d80fbcaca266696a7d6de7342876d104c42e91e9)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>